Release of 1.9.10
Redundancy 1.9.10 Changelog
===========================
Bugfixes
============
- Re-Added the branding at the registration
- Fixed a bug when a directory was renamed and the name existed already
- Fixed a bug which allowed to create directories with empty names
- Fixed a bug which redirected the user to a non existant directory when
the program should be installed
- Fixed a bug which blocked the playing of a multimedia file
- Fixed a bug caused a strange action of the administration panel (issue
#15)
- Fixed a bug which blocked the usage of the administration panel (issue
#16)
- Fixed a bug giving wrong feedback at the file upload (issue #17)
- Fixed a bug connected to issue #17 at the file upload
- Fixed a translation bug (issue #14)
- Fixed a bug to get an empty message when the get parameter "message"
is incorrect
- Fixed a bug causing a wrong download filename (issue #20)
- Fixed a bug allowing unlimited filename length (added
Program_FileSystem_Name_Max_Length to control it)
- Fixed a bug causing an error message in the webserver's log files
containing the variable Program_Dir is not known
Security Bugfixes
=================
- Fixed a bug which allowed it to see the properties of a file if the
hash was known<br>
Feature additions
=================
- Added better feedback at the user registration
- Added better feedback if registration is disabled
- Applied a better stylesheet for the context menu
- Register and recover links are now only displayed if the option is
enabled
- The filename or directory name is now displayed at the delete confirm
dialog (issue #12)
- Users can be created with the administration panel (issue #13)
- Users can be enabled or disabled with the administration panel
- Added a list of last changed files and folders (issue #8)
- Added a new bootstrap based theme
- Added a mobile view
- Added setting "use_buffer" to use the object buffer of php
- Added setting "Program_Share_Link_Length" to control the length of
share links (default: 7 chars)
Features removed
================
- Removed support vor any theme before 1.9.10

+++ Includes/rename.inc.php
@@ -25,7 +25,7 @@
 	//start a session if needed
 	if (isset($_SESSION) == false)
 			session_start();
-			
+	$success = false;
 	//only proceed if a post parameter is set
 	if ($_SESSION["role"] != 3  )
 	{		
@@ -38,14 +38,21 @@
 					$source = mysqli_real_escape_string($connect,$_GET["source"]);
 					$target = mysqli_real_escape_string($connect,$_POST["newname"]);
 					$old_root= mysqli_real_escape_string($connect,$_GET["old_root"]);
-					createDir($old_root,$target);
-					$Dir_ID = getDirectoryID($source);
-					//TODO: DIR renmae
-					//Step 1 create new dir (new name) - check
-					moveContents($source,$_SESSION["currentdir"].$target);
-					//Step 2 move contents -check					
-					//STep 3 delete old dir
-					deleteDir($source);
+					if (fs_file_exists($target,$old_root) == false && strlen($target) <= $GLOBALS["config"]["Program_FileSystem_Name_Max_Length"]){
+						createDir($old_root,$target);
+						$Dir_ID = getDirectoryID($source);
+						//TODO: DIR renmae
+						//Step 1 create new dir (new name) - check
+						moveContents($source,$_SESSION["currentdir"].$target);
+						//Step 2 move contents -check					
+						//STep 3 delete old dir
+						deleteDir($source);
+						$success = true;
+					}
+					else
+					{
+						$success = false;
+					}
 				}
 				else{
 					//include the dataBase file
@@ -55,8 +62,10 @@
 						$hash = mysqli_real_escape_string($connect,$_GET["file"]);
 					else
 						$hash = mysqli_real_escape_string($connect,$_POST["file"]);
-					if (fs_file_exists($newname,$_SESSION["currentdir"]) == false) 
+					if (fs_file_exists($newname,$_SESSION["currentdir"]) == false && strlen($newname) <= $GLOBALS["config"]["Program_FileSystem_Name_Max_Length"]) {
 						$insert = mysqli_query($connect,"Update Files Set Displayname='$newname' where Hash ='$hash'") or die("Error: 017 ".mysqli_error($connect));	
+						$success = true;
+					}				
 				}
 				if (isset($_POST["api_key"]))
 				{		
@@ -64,10 +73,16 @@
 					exit;		
 				}
 				else{	
-					if ($GLOBALS["config"]["Program_Debug"] != 1){
-						header("Location: ./index.php?module=list&dir=".$_SESSION["currentdir"]);
-						exit;
+					if ($GLOBALS["config"]["Program_Debug"] != 1){						
+						if ($success == true){
+							header("Location: ./index.php?module=list&dir=".$_SESSION["currentdir"]."&message=rename_success");
+						}
+						else
+						{
+							header("Location: ./index.php?module=list&dir=".$_SESSION["currentdir"]."&message=rename_fail");
+						}
 					}	
+					
 				}
 			}		
 		}		
@@ -86,6 +101,6 @@
 	}	
 ?>
 <form method="POST" action="index.php?module=rename<?php echo $suffix;?>" align = "center">
-<div class = 'contentWrapper'><tag><?php echo sprintf($GLOBALS["Program_Language"]["Rename"],$file);?> <r></tag><input name="newname">
-<input type=submit name=submit value="<?php echo $GLOBALS["Program_Language"]["Rename_Button"];?>"></div>
+<tag><?php echo sprintf($GLOBALS["Program_Language"]["Rename"],$file);?> <r></tag><input name="newname">
+<input class = "btn btn-default" type=submit name=submit value="<?php echo $GLOBALS["Program_Language"]["Rename_Button"];?>">
 </form>
